[ContentType] First pass at refactoring content type.
Now, only default_path will be kept in the model - the rest is kept in app.yml. The content type will be abstracted to an object, and when you think of a "content type", you should think about the configuration in app.yml for it, not the model. The model will be handled in the background for you, and only exists so that any content-based information on the content type can still live in the database (like default_path, theme [later], etc).

The big advantage, in my opinion, is that you as a developer won't need to worry about a content type record - you just need to worry about defining how a content type model will be rendered via app.yml.