Adding new decoupled sfSympalContentTypeObject and sfSympalContentObject classes that put the logic away from the model.
I really likke this for sfSympalContentTypeObject as it joins configuration with some content. sfSympalContentObject, however, may not be necessary, as it's defined in the database first and foremost. It was abstracted because it needs its sfSympalContentTypeObject in order to get accurate information about its type. If I cheat and use sfApplicationConfiguration::getActive(), I could get around this by retrieving the sfSympalContentTypeObject from the model (sfSympalContentType).