Removing the content type object - an unnecessary abstraction.
The content type is a non-database object primarily. We join it to the database for a little bit of data (whereas in Doctrine 2, this would be just one object, where only a few fields persist). The sfSympalContent object, however, is primarily database driven - there is no config side of it. So, it's ok as JUST a Doctrine record.