Having the slug and content_slug on sfSympalContent, while everything else was on the content type model was very awkward. Dealing with the $record->Content and $content->Record relationship in both directions was causing all sorts of problems. This is the first step to remove slug and i18n_slug from sfSympalContent.
The importance of slug is primarily in url creation. Now, you can have the slug on your content type model and choose to have it in your url or not (or make up a different field name that acts like a slug). My hope is that with good awareness of i18n, the i18n_slug functionality can be recreated by having some sort of slug field that is i18n'ed and then included in your route.